requirements.txt
discord.py==2.1.0
python-dotenv==0.20.0
interaction==1.3
PyYAML==6.0
tls-client==0.1.8
websockets==10.4
typing==3.7.4.3
EdgeGPT==0.0.48